chore(CLICommand): Separates the commands into its own files PE-342
